Deprecate ContainerApp
Instead of asking developers to extend Application for storing the global shell,
libsu is updated to have the ability to inject the root shell directly into
the Application's context. This auto injection feature is used by default
unless the developer explicitly override the Container with a new instance.

This means we do not need ContainerApp anymore. Deprecate it so developers
no longer uses the useless class.